Initial impl of init command
Refactoring pending - how to handle common arguments that should be
optional to the user, e.g. username, keypair location?

We don't want to force the user to enter these every time, but I want
to avoid cluttering the code with complicated default logic. Perhaps I
could provide a "context map" argument? Reader monad?

Note that some of the options are sensitive (e.g. secret key) so I
can't just read them all at startup. Options should be as functions, so
they are accessed on-demand and their results are not stored.